Continue trying to implement unit testing.
Merged the AgentInstaller from feature/201808200214-FirstUnitTests.

Creating a JAR with dependencies out of the agent project.

The JAR from the agent project also has the agent jar specified in the
manifest.

Test still not working. Need to somehow load the jar with dependencies
as the maven dependency into the unit test project.